requirements.txt
fastapi
uvicorn
pytest
httpx
authlib
itsdangerous
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
